3496 - Move StatisticsTable to StatisticsSidePanel
