+++ src/client/pages/Geo/StatisticsSidePanel/StatisticsTable/StatisticsTable.scss
