+++ src/client/pages/Geo/StatisticsSidePanel/StatisticsTable/index.ts
